gave up hash features.

+++ statusSearch.h
@@ -8,7 +8,6 @@
 #include "icld.h"
 #include "boardEvaluate.h"
 #include "mathFunc.h"
-#include "Zoribrist_hash.h"
 #include "customTypes.h"
 
 void generate_possible_pos(drop_choice *drop_choice1, int *num, int search_player_side);
@@ -24,7 +23,7 @@ void search_init();
 void time_init();
 
 //standarise the grade to fit with the scale of 8^remaining_depth
-int64_t grade_standardise(int64_t grade);
+long long int grade_standardise(long long int grade);
 
 void time_display(uint32_t depth, long long int dfs_grade);
